requirement.txt
numpy==1.26.1
pyasn1==0.5.0        
pycryptodomex==3.19.0
rsa==4.9